+++ src/inspector/types/commontypes.h
@@ -56,7 +56,7 @@ struct ElementKey
 using ElementKeyList = QList<ElementKey>;
 using ElementKeySet = QSet<ElementKey>;
 
-inline uint qHash(const ElementKey& key)
+inline size_t qHash(const ElementKey& key)
 {
     QString subtypePart = key.subtype >= 0 ? QString::number(key.subtype) : "";
     return qHash(QString::number(static_cast<int>(key.type)) + subtypePart);
